Add random color
Added random color instead of white and tried to add the darkening
effect but without success. After many hours decided to move on.
Will be back to it later.
Same with random color for every grid you hover over. Managed to do it,
but it made all the palette unusable. Decided to turn it off.
Later.